chore(frontend-test): implement test for actions,reducers and components [#Delivers 153048798 ]

+++ client/src/components/group/Group.jsx
@@ -1,10 +1,12 @@
 import React, { Component } from 'react';
+import swal from 'sweetalert';
 import SingleGroupMessage from '../message/SingleGroupMessage.jsx';
 import GroupHeader from './GroupHeader.jsx';
 import GroupSideBar from '../group/GroupSideBar.jsx';
 import Loader from '../loaders/Loader.jsx';
 import { searchUsersNotInGroup } from '../../utils/postItApi';
 import NewMessage from '../message/NewMessage.jsx';
+import { getAuthUser } from '../../utils/authservice';
 
 /**
  * @class
@@ -49,6 +51,7 @@ class Group extends Component {
     $('.modal').modal({
       opacity: 0.5,
     });
+
     $('.dropdown-button').dropdown({
       inDuration: 300,
       outDuration: 225,
@@ -65,7 +68,19 @@ class Group extends Component {
    */
   handleRemoveMember(userId, userIndex) {
     const groupId = this.props.params.id;
-    this.props.actions.groupActions.deleteGroupMember(groupId, userId, userIndex);
+
+    swal({
+      title: 'Are you sure?',
+      text: 'You are about to remove this group member. This action cannot be undone',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((removeMember) => {
+        if (removeMember) {
+          this.props.actions.groupActions.deleteGroupMember(groupId, userId, userIndex);
+        }
+      });
   }
   /**
    * @return {state} state
@@ -87,7 +102,18 @@ class Group extends Component {
     const groupIndex = this.props.groups.findIndex(group => group.id ==
       id);
 
-    this.props.actions.groupActions.deleteGroup(id, groupIndex);
+    swal({
+      title: 'Are you sure?',
+      text: 'Once you delete this group. You cannot undo this action',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((deleteGroup) => {
+        if (deleteGroup) {
+          this.props.actions.groupActions.deleteGroup(id, groupIndex);
+        }
+      });
   }
   /**
    * @returns { group } group
@@ -97,7 +123,18 @@ class Group extends Component {
     const groupIndex = this.props.groups.findIndex(group => group.id ==
       id);
 
-    this.props.actions.groupActions.leaveGroup(id, groupIndex);
+    swal({
+      title: 'Are you sure?',
+      text: 'Once you leave this group. You cannot undo this action',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((leaveGroup) => {
+        if (leaveGroup) {
+          this.props.actions.groupActions.leaveGroup(id, groupIndex);
+        }
+      });
   }
   /**
    * @param { bool } bool
@@ -153,60 +190,61 @@ class Group extends Component {
 
     const group = this.props.groups[index];
     const { messages, messageIsLoading } = this.props;
+
     return (
       <div>
         <div className="row">
             <div className="col s12 m4 l3">
-                  <GroupSideBar
-                      groupUsers = {this.props.groupUsers}
-                      search = {this.state.search}
-                      onSearchChange={this.onSearchChange}
-                      searchLoading ={this.state.searchLoading}
-                      searchResult={this.state.searchResults}
-                      searchErrorMessage={this.state.searchErrorMessage}
-                      onAddUser={this.onAddUser}
-                      onDeleteGroup={ this.onDeleteGroup }
-                      showDelete ={ this.state.showDelete }
-                      onMouseEnterDelete={this.onMouseEnterDelete}
-                      onMouseLeaveDelete={this.onMouseLeaveDelete}
-                      handleRemoveMember={this.handleRemoveMember}
-                      group={group}/>
-                </div>
+              <GroupSideBar
+                groupUsers = {this.props.groupUsers}
+                search = {this.state.search}
+                onSearchChange={this.onSearchChange}
+                searchLoading ={this.state.searchLoading}
+                searchResult={this.state.searchResults}
+                searchErrorMessage={this.state.searchErrorMessage}
+                onAddUser={this.onAddUser}
+                onDeleteGroup={ this.onDeleteGroup }
+                showDelete ={ this.state.showDelete }
+                onMouseEnterDelete={this.onMouseEnterDelete}
+                onMouseLeaveDelete={this.onMouseLeaveDelete}
+                handleRemoveMember={this.handleRemoveMember}
+                authUser = { getAuthUser() }
+                />
+              </div>
             <div className="col s12 m8 l9">
               {
                 (messageIsLoading) ? (<Loader/>) : ('')
               }
                 <GroupHeader
-                    headerText={ group.name }
-                    onExitGroup={ this.onExitGroup }
-                    onDeleteGroup={ this.onDeleteGroup }/>
+                  headerText={ group ? group.name : '' }
+                  onExitGroup={ this.onExitGroup }
+                  onDeleteGroup={ this.onDeleteGroup }/>
+
                 <div id="messages" className="messageOverflow">
-                    <ul>
-                        {
-                            messages.map((message, i) =>
-                                <SingleGroupMessage message={message}
-                                key={i} i={i}/>)
-                        }
-                          {
-                            (messages.length === 0) ?
-                            (<div className="col s6 offset-s2 not-found center-align">
-                                <h5>No messages in this group</h5>
-                                <p>Post a message now</p>
-                            </div>)
-                            : ''
-                        }
-                    </ul>
+                  <ul>
+                    {
+                      messages.map((message, i) =>
+                        <SingleGroupMessage message={message}
+                        key={i} i={i}/>)
+                    }
+                    {
+                      (messages.length === 0) ?
+                      (<div className="col s6 offset-s2 not-found center-align">
+                          <h5>No messages in this group</h5>
+                          <p>Post a message now</p>
+                      </div>)
+                      : ''
+                    }
+                  </ul>
                 </div>
 
-                  <div className="messaging-area">
-                    <NewMessage
-                      groupId={this.props.params.id}
-                      {...this.props}/>
-                  </div>
+                <div className="messaging-area">
+                  <NewMessage
+                    groupId={this.props.params.id}
+                    {...this.props}/>
+                </div>
 
             </div>
-            
-
         </div>
 
       </div>
